pulled changesInBaseName and designationIsFixed (and related getters and methods) up out of grey-box.vue into analyze-results.vue (parrent) because the result of these getters was needed after all instances of grey-box.vue have been destroyed (per changes in 5664 redesign) / that's where they should have been to begin with
