+++ client/src/vue.d.ts
@@ -1,4 +1,5 @@
 import Vue from 'vue-property-decorator'
+import quill from 'quill'
 
 export interface DesignationI {
   words: string[]
