+++ client/tests/unit/analyze-results.spec.ts
@@ -89,7 +89,7 @@ describe('AnalyzeResults', () => {
           done()
         })
         test('You may proceed message is displayed', () => {
-          expect(wrapper.text()).toContain('You may proceed')
+          expect(wrapper.text()).toContain('You May Proceed')
         })
         test('Corrected name generated by front-end is correct', () => {
           expect(wrapper.vm.name).toBe(data.corrected)
